# docs/contracts/vault-openbao-compat.openapi.json
{
  "components": {
    "schemas": {
      "VaultErrorResponse": {
        "properties": {
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "errors"
        ],
        "type": "object"
      },
...
      "VaultPKIIssueRequest": {
        "properties": {
          "common_name": {
            "type": "string"
          },
          "ttl": {
            "type": "string"
          },
          "ttl_seconds": {
            "type": "integer"
          }
        },
        "required": [
          "common_name"
        ],
        "type": "object"
      },
      "VaultPKIIssueResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/VaultPKIIssueData"
          },
          "lease_duration": {
            "type": "integer"
          },
          "lease_id": {
            "type": "string"
          },
          "renewable": {
            "type": "boolean"
          },
          "request_id": {
            "type": "string"
          },
          "warnings": {}
        },
        "required": [
          "request_id",
          "lease_id",
          "renewable",
          "lease_duration",
          "data",
          "warnings"
        ],
        "type": "object"
      },
      "VaultTokenLookupSelf": {
        "properties": {
          "accessor": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "entity_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "meta": {
            "properties": {
              "tenant_id": {
                "type": "string"
              }
            },
            "required": [
              "tenant_id"
            ],
            "type": "object"
          },
          "num_uses": {
            "type": "integer"
          },
          "orphan": {
            "type": "boolean"
          },
          "path": {
            "type": "string"
          },
          "policies": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "renewable": {
            "type": "boolean"
          },
          "ttl": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "accessor",
          "display_name",
          "entity_id",
          "meta",
          "num_uses",
          "orphan",
          "path",
          "policies",
          "renewable",
          "ttl"
        ],
        "type": "object"
      },
      "VaultTokenLookupSelfResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/VaultTokenLookupSelf"
          },
          "lease_duration": {
            "type": "integer"
          },
          "lease_id": {
            "type": "string"
          },
          "renewable": {
            "type": "boolean"
          },
          "request_id": {
            "type": "string"
          },
          "warnings": {}
        },
        "required": [
          "request_id",
          "lease_id",
          "renewable",
          "lease_duration",
          "data",
          "warnings"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "VaultToken": {
        "description": "A normal trstctl tenant API token carried in the header used by the stock Vault CLI.",
        "in": "header",
        "name": "X-Vault-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Compatibility subset for stock Vault/OpenBao migration clients. The native /api/v1 OpenAPI document remains the complete trstctl API contract.",
    "title": "trstctl Vault/OpenBao Compatibility API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/auth/token/lookup-self": {
      "get": {
        "operationId": "vaultCompatTokenLookupSelf",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultTokenLookupSelfResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible success response."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          }
        },
        "security": [
          {
            "VaultToken": []
          }
        ],
        "summary": "Return Vault/OpenBao-shaped metadata for the trstctl API token"
      },
      "post": {
        "operationId": "vaultCompatTokenLookupSelfPost",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultTokenLookupSelfResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible success response."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          }
        },
        "security": [
          {
            "VaultToken": []
          }
        ],
        "summary": "Return Vault/OpenBao-shaped metadata for the trstctl API token"
      }
    },
    "/v1/pki/issue/{role}": {
      "post": {
        "operationId": "vaultCompatPKIIssue",
        "parameters": [
          {
            "description": "Vault PKI role label accepted for CLI compatibility; trstctl maps issuance to the requested common name.",
            "in": "path",
            "name": "role",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional for stock Vault/OpenBao clients. When omitted, trstctl derives a deterministic key from method, path, and body so retries do not duplicate mutations.",
            "in": "header",
            "name": "Idempotency-Key",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaultPKIIssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultPKIIssueResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible success response."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          }
        },
        "security": [
          {
            "VaultToken": []
          }
        ],
        "summary": "Issue a Vault/OpenBao-shaped short-lived certificate and private key",
        "x-trstctl-permission": "secrets:write",
        "x-trstctl-sensitive-response": true
      },
      "put": {
        "operationId": "vaultCompatPKIIssuePut",
        "parameters": [
          {
            "description": "Vault PKI role label accepted for CLI compatibility; trstctl maps issuance to the requested common name.",
            "in": "path",
            "name": "role",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional for stock Vault/OpenBao clients. When omitted, trstctl derives a deterministic key from method, path, and body so retries do not duplicate mutations.",
            "in": "header",
            "name": "Idempotency-Key",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaultPKIIssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultPKIIssueResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible success response."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultErrorResponse"
                }
              }
            },
            "description": "Vault/OpenBao-compatible error envelope."
          }
        },
        "security": [
          {
            "VaultToken": []
          }
        ],
        "summary": "Issue a Vault/OpenBao-shaped short-lived certificate and private key",
        "x-trstctl-permission": "secrets:write",
        "x-trstctl-sensitive-response": true
      }
    },
...
  }
}